requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic[email]
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.5
aiomysql==0.2.0
requests==2.31.0  # For testing API endpoints 